DB Servere Goşuldy!

+++ CV_Proyekt/Models/SqlModels/Skill.cs
@@ -2,9 +2,9 @@
 {
     public class Skill
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string Name { get; set; }
         public int Percentage { get; set; }
-        public string Type { get; set; } 
+        public string Type { get; set; }
     }
 }
